modularize matOptimize for larch
