+++ CMakeLists.txt
@@ -135,6 +135,7 @@ if(DEFINED Protobuf_PATH)
         src/matOptimize/optimize_tree.cpp
         src/matOptimize/import_vcf_fast.cpp
         src/matOptimize/condense.cpp
+        src/matOptimize/optimize_inner_loop.cpp
         src/matOptimize/VCF_load_tree.cpp
         src/matOptimize/main_load_tree.cpp
         src/matOptimize/main_helper.cpp
@@ -321,6 +322,7 @@ else()
         src/matOptimize/mutation_annotated_tree_load_store.cpp
         src/matOptimize/mutation_annotated_tree_nuc_util.cpp
         src/matOptimize/Mutation_Collection.cpp
+        src/matOptimize/reassign_states.cpp
         src/matOptimize/check_samples.cpp
         src/matOptimize/optimize_tree.cpp
         src/matOptimize/condense.cpp
@@ -374,6 +376,8 @@ else()
         src/matOptimize/main_helper.cpp
         src/matOptimize/Mutation_Collection.cpp
         src/matOptimize/Fitch_Sankoff.cpp
+        src/matOptimize/optimize_inner_loop.cpp
+        src/matOptimize/reassign_states.cpp
         src/matOptimize/check_samples.cpp
         #src/matOptimize/priority_conflict_resolver_cross_only.cpp
         src/matOptimize/priority_conflict_resolver.cpp
