+++ src/matOptimize/main.cpp
@@ -60,10 +60,6 @@ void log_flush(int) {
 namespace po = boost::program_options;
 namespace MAT = Mutation_Annotated_Tree;
 MAT::Node* get_LCA(MAT::Node* src,MAT::Node* dst);
-static void make_output_path(std::string& path_template) {
-    auto fd=mkstemps(const_cast<char*>(path_template.c_str()),3);
-    close(fd);
-}
 static void load_reference(std::string fasta_fname){
     auto fh=fopen(fasta_fname.c_str(), "r");
     char* seq_name=nullptr;
@@ -95,9 +91,6 @@ void print_file_info(std::string info_msg,std::string error_msg,const std::strin
         exit(EXIT_FAILURE);
     }
 }
-#define DRIFT_MASK 0x80000000
-#define ALL_DIR_MASK 0x40000000
-#define RADIUS_MASK 0x3fffffff
 int main(int argc, char **argv) {
     int ignored;
     auto init_result=MPI_Init_thread(&argc, &argv,MPI_THREAD_MULTIPLE,&ignored);
@@ -241,9 +234,7 @@ int main(int argc, char **argv) {
         } else if (input_nh_path!=""&&input_vcf_path!="") {
             print_file_info("Load starting tree from", "starting tree file,-t", input_nh_path);
             print_file_info("Load sample variant from", "sample vcf,-v", input_vcf_path);
-        } else if (input_nh_path!=""&&diff_file_path!="") {
-        
-        } 
+        }
         else {
             fputs("Input file not completely specified. Please either \n"
                   "1. Specify an intermediate protobuf from last run with -a to continue optimization, or\n"
@@ -275,7 +266,6 @@ int main(int argc, char **argv) {
         fprintf(stderr,"Run kill -s SIGUSR2 %d to apply all the move found immediately, then output and exit.\n",pid);
         fprintf(stderr,"Using %d threads. \n",num_threads);
         auto start_time=std::chrono::steady_clock::now();
-        auto save_period=std::chrono::minutes(minutes_between_save);
         fprintf(stderr, "Will drift for %d iterations \n",drift_iterations);
 
         //Loading tree
@@ -349,7 +339,6 @@ int main(int argc, char **argv) {
             save_final_tree(t, output_path);
             return 0;
         }
-        auto last_save_time=std::chrono::steady_clock::now();
         size_t new_score;
         size_t score_before;
         int stalled = -1;
@@ -397,87 +386,12 @@ int main(int argc, char **argv) {
             if (nodes_to_search.empty()) {
                 break;
             }
-            bool isfirst_this_iter=true;
             //Actual optimization loop
-            while (!nodes_to_search.empty()) {
-                auto dfs_ordered_nodes=t.depth_first_expansion();
-                std::mt19937_64 rng;
-                std::shuffle(nodes_to_search.begin(), nodes_to_search.end(),rng);
-                bool distribute=(process_count>1)&&(nodes_to_search.size()>1000);
-                if (distribute) {
-                    MPI_Request req;
-                    int radius_to_boardcast=abs(radius);
-                    if (allow_drift) {
-                        radius_to_boardcast|=DRIFT_MASK;
-                    }
-                    if(search_all_dir) {
-                        radius_to_boardcast|=ALL_DIR_MASK;
-                        fprintf(stderr, "Search all directions\n");
-                    }
-                    MPI_Ibcast(&radius_to_boardcast, 1, MPI_INT, 0, MPI_COMM_WORLD, &req);
-                    fprintf(stderr, "Sent radius\n");
-                    MPI_Wait(&req, MPI_STATUS_IGNORE);
-                    fprintf(stderr, "Start Send tree\n");
-                    t.MPI_send_tree();
-                }
-                adjust_all(t);
-                use_bound=true;
-                std::vector<size_t> nodes_to_search_idx;
-                nodes_to_search_idx.reserve(nodes_to_search.size());
-                for(const auto node:nodes_to_search) {
-                    nodes_to_search_idx.push_back(node->dfs_index);
-                }
-                std::vector<size_t> defered_nodes;
-                auto next_save_time=minutes_between_save?last_save_time+save_period:std::chrono::steady_clock::time_point::max();
-                bool do_continue=true;
-                auto search_stop_time=next_save_time;
-                if (no_write_intermediate||search_end_time<next_save_time) {
-                    search_stop_time=search_end_time;
-                }
-                optimize_tree_main_thread(nodes_to_search_idx, t,std::abs(radius),movalbe_src_log,allow_drift,log_moves?iteration:-1,defered_nodes,distribute,search_stop_time,do_continue,search_all_dir,isfirst_this_iter
+            new_score=optimize_inner_loop(nodes_to_search,t,radius,
 #ifdef CHECK_STATE_REASSIGN
-                                          , origin_states
+            origin_states,
 #endif
-                                         );
-                isfirst_this_iter=false;
-                fprintf(stderr, "Defered %zu nodes\n",defered_nodes.size());
-                nodes_to_search.reserve(defered_nodes.size());
-                nodes_to_search.clear();
-                for (auto idx : defered_nodes) {
-                    if (t.get_node(idx)) {
-                        nodes_to_search.push_back(t.get_node(idx));
-                    }
-                }
-                auto curr_score=t.get_parsimony_score();
-                if(curr_score>=new_score) {
-                    nodes_to_search.clear();
-                }
-                new_score=curr_score;
-                fprintf(stderr, "parsimony score after optimizing: %zu,with radius %d, second from start %ld \n\n",
-                        new_score,std::abs(radius),std::chrono::duration_cast<std::chrono::seconds>(std::chrono::steady_clock::now()-start_time).count());
-                if(!no_write_intermediate) {
-                    intermediate_writing=intermediate_template;
-                    make_output_path(intermediate_writing);
-                    auto save_start=std::chrono::steady_clock::now();
-                    t.save_detailed_mutations(intermediate_writing);
-                    rename(intermediate_writing.c_str(), intermediate_pb_base_name.c_str());
-                    last_save_time=std::chrono::steady_clock::now();
-                    fprintf(stderr, "Took %ldsecond to save intermediate protobuf\n",std::chrono::duration_cast<std::chrono::seconds>(last_save_time-save_start).count());
-                }
-                if(allow_drift) {
-                    MAT::save_mutation_annotated_tree(t, intermediate_nwk_out+std::to_string(iteration)+".pb.gz");
-                }
-                if (std::chrono::steady_clock::now()>=search_end_time) {
-                    break;
-                }
-                if (interrupted) {
-                    break;
-                }
-                if (allow_drift) {
-                    nodes_to_search.clear();
-                }
-                search_all_dir=true;
-            }
+            allow_drift,search_all_dir,minutes_between_save,no_write_intermediate,search_end_time,start_time,log_moves,iteration,intermediate_template,intermediate_pb_base_name,intermediate_nwk_out);
             if (interrupted) {
                 break;
             }
